watchlist rendering but working on props
